Refactor code formatter and make it more maintainable
