+++ src/main/kotlin/io.github.divinenickname.kotlin.utgen.idea.plugin/GenerateUnitTest.kt
@@ -10,14 +10,9 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
 import com.intellij.psi.PsiManager
-import com.intellij.psi.codeStyle.CodeStyleManager
 import com.squareup.kotlinpoet.FileSpec
 import io.github.divinenickname.kotlin.utgen.core.UnitTestGenerator
 import org.jetbrains.kotlin.idea.KotlinFileType
-import org.jetbrains.kotlin.lexer.KtTokens
-import org.jetbrains.kotlin.psi.KtClassOrObject
-import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.psi.KtFunction
 import java.io.File
 
 class GenerateUnitTest : AnAction() {
@@ -39,27 +34,12 @@ class GenerateUnitTest : AnAction() {
 
         PsiManager.getInstance(project).findDirectory(directory!!)!!
             .takeIf { it.findFile(psiFile.name) == null }
-            ?.run {
-                ApplicationManager.getApplication().runWriteAction { this.add(psiFile) }
-            }
+            ?.run { ApplicationManager.getApplication().runWriteAction { this.add(psiFile) } }
     }
 
-    private fun makePsiFile(project: Project, fileSpec: FileSpec): PsiFile {
-        val fileName = "${fileSpec.name}.kt"
-        val psiFile =  PsiFileFactory.getInstance(project)
-            .createFileFromText(fileName, KotlinFileType.INSTANCE, fileSpec.toString())
-
-        (psiFile as KtFile).declarations.forEach { declaration ->
-            if (declaration is KtClassOrObject) {
-                declaration.declarations.filterIsInstance<KtFunction>()
-                    .filter { it.hasModifier(KtTokens.PUBLIC_KEYWORD) }
-                    .forEach { it.removeModifier(KtTokens.PUBLIC_KEYWORD) }
-            }
-        }
-        CodeStyleManager.getInstance(project).reformat(psiFile)
-
-        return psiFile
-    }
+    private fun makePsiFile(project: Project, fileSpec: FileSpec): PsiFile = PsiFileFactory.getInstance(project)
+        .createFileFromText("${fileSpec.name}.kt", KotlinFileType.INSTANCE, fileSpec.toString())
+        .apply { KotlinCodeFormatter(project, this).apply() }
 
     override fun update(e: AnActionEvent) {
         val presentation = e.presentation
